+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## [4.8.1](https://github.com/adobe/helix-dispatch/compare/v4.8.0...v4.8.1) (2021-04-08)
+
+
+### Bug Fixes
+
+* clean up x-error values ([d2e3c04](https://github.com/adobe/helix-dispatch/commit/d2e3c04272f6114ca1fcca12a045b6e22014dc98)), closes [adobe/helix-deploy#201](https://github.com/adobe/helix-deploy/issues/201)
+
+# [4.8.0](https://github.com/adobe/helix-dispatch/compare/v4.7.2...v4.8.0) (2021-04-06)
+
+
+### Bug Fixes
+
+* **index:** truncate logged URL effectively ([973f662](https://github.com/adobe/helix-dispatch/commit/973f6629a4bae8f9928c02f51ac972732b0fa961))
+* **logging:** truncate URL in logs ([9380a21](https://github.com/adobe/helix-dispatch/commit/9380a21f323afdb47426dfb2406b1a400256b0bc))
+
+
+### Features
+
+* **deploy:** deploy to google ([9aedd85](https://github.com/adobe/helix-dispatch/commit/9aedd85ba3c1a3bcc025a716e3c54aa3e4f51c19))
+* **google:** extract function execution id from google ([cebba0b](https://github.com/adobe/helix-dispatch/commit/cebba0b234a6a2b9e863606a9b329927603245f6))
+
+## [4.7.2](https://github.com/adobe/helix-dispatch/compare/v4.7.1...v4.7.2) (2021-04-05)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-status to v9.1.5 ([#476](https://github.com/adobe/helix-dispatch/issues/476)) ([ae92d9e](https://github.com/adobe/helix-dispatch/commit/ae92d9eab7c827d5b34f53409c0f2248ff672d00))
+
+## [4.7.1](https://github.com/adobe/helix-dispatch/compare/v4.7.0...v4.7.1) (2021-04-03)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-status to v9.1.4 ([#475](https://github.com/adobe/helix-dispatch/issues/475)) ([433dd31](https://github.com/adobe/helix-dispatch/commit/433dd319b1e377e1e7269fb83046671efbe01381))
+
 # [4.7.0](https://github.com/adobe/helix-dispatch/compare/v4.6.13...v4.7.0) (2021-03-30)